CactusTV/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
neo4j==5.15.0
numpy==1.26.2
protobuf==4.21.12
pymongo==4.6.1
pytz==2023.3.post1
Werkzeug==3.0.1
